daltonizer: "error steering"
Previously, when correcting for colorblindness, the error was blindly
splattered across the other cones in hopes that visibility would
improve. But what if we directed it towards a specific color?

My guess is that by being intentional about where error comes from and
how it's distributed, it's possible to create a filter that corrects
more and distorts less. Lots more experimentation is needed.