effects: set up a bunch of filter variations
In an upstream issue, someone asked for filters that are more
intentional about how they affect color. Since the "error steering"
filters already do this, it's easy to set up a bunch of variations on
the theme that they asked for. Let's see what people like, I guess.

It hurts me to be so un-clever about initializing all the filter
options, but sometimes, it's nice to be able to see the actual data
being declared instead of a bunch of nested loops. Let's see how long I
can leave it alone before I bring the loops back...

Also, cut down on the number of properties in filter definitions. I like
the "tritan hack" and it's my extension, so it's the only option. I like
the M_HPE transform better, so it's the only option.