opponent: clean up rgb2opp preconditioning
Fixes sim_rgb2lms so it uses cone sensitivities instead of primary
intensities, though the difference is minor. No wonder I was able to
miss it for so long.

Also changes the way that lms2opp is normalized to better match luma and
to avoid tinting yellow into green when correcting tritanomaly.