opponent: add trivial simulation filters
These use all the same matrix setup magic as the correction filters, but
don't need the same complex shader. They're mostly intended as a way to
visualize what the OCS filters are trying to correct for, but if they
eventually evolve into useful filters in their own right, so much the
better.