extension: always attach effect to a Clutter.Clone
What a nightmare.

Attaching to a Clutter.Clone instead of to uiGroup directly fools
Clutter into turning off some "optimizations" that cause misbehavior.
This is a workaround suggested in a very old mutter merge request about
effectively this exact issue:

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2269

Unfortunately, the screen magnifier tool creates its own UI over the
top of uiGroup, which will obscure our Clone. Since these are
complementary accessibility features, it would be nice to allow them to
work together. So, watch uiGroup for the screen magnifier popping up and
attach to it instead.

One complication: the most obvious notification that the screen
magnifier is disabling is when the 'destroy' signal fires on the
Clutter.Actor that our Clone was attached to. Anything directly attached
to it at that point becomes radioactive and can't safely be reused
without possibly crashing gnome-shell. I think all of that has been
worked out at this point.