opponent: compute transforms at the same time
Since the simulated transform needs parts of the ideal transform for
basically every conditioning step, it makes sense to just compute them
together.

This change was motivated by yet another change to conditioning that
depends (at least on paper) on the ideal transform: offsetting chroma
only for the simulated transform, and much more importantly, making the
offsets proportional to simulated luma sensitivity. This fixes the
weakness of the simulation filter, which should improve correction
quality.