opponent: ignore RGB value difference
The RGB weights were low enough not to have a significant impact on the
final value (or deal with the tendency of the tritanomaly filter to
produce extreme blue component values). Raising them makes the final
color look worse, so apparently there's no reason to bother with them at
all. Huh.

The developer is slightly concerned about this development and about
whether or not they even understand how this filter works.