opponent: use brightness as-is
The developer is unsure what has changed to allow for this, but the
original decorrelation given by Wandell (and similarly with that of
Ingling and Tsou) now gives good correction results when it previously
did not. This feels much less hacky, and as a bonus, it causes the
protanopia correction to actually do something and reduces the
purple/green tint caused by the tritanopia correction.