opponent: more general chroma reduction formula
As it turns out, the previous formula to find a chroma scaling factor
was a special case of a more general formula. The more general version
behaves correctly when the simulated color blindness chroma range is
less than half the size of the trichromat range. Previously, that would
cause colors to saturate (which was intentional--otherwise, chroma
wouldn't have been monotonic!).

The previous formula was equivalent to this one with k fixed at 0.5.
Because the new formula always uses an optimal value of k, the resulting
chroma values track more closely with the input value. That should lead
to slightly more vibrant colors.

This looks convoluted, but has nice properties that other functions
don't. It's surprisingly difficult to find a function with a slope of 1
at zero chroma (so it tracks the trichromat chroma well for light
colors) that has the right start and end value and is monotonic in
between.